Construída requisição Delete
